refactor: isolate lora profile mutation sync
